Extract `useActiveElement`
