+++ apps/lite/ui/src/routes/project/$id/ProjectPreviewLayout.tsx
@@ -1,4 +1,4 @@
-import { FC, ReactNode, useRef, useSyncExternalStore } from "react";
+import { FC, ReactNode, useRef } from "react";
 import { Group, Panel, Separator, useDefaultLayout } from "react-resizable-panels";
 import {
 	isPanelVisible,
@@ -10,23 +10,14 @@ import { useAppSelector } from "#ui/state/hooks.ts";
 import styles from "./ProjectPreviewLayout.module.css";
 import { useMergedRefs } from "@base-ui/utils/useMergedRefs";
 import { classes } from "#ui/classes.ts";
+import { useActiveElement } from "#ui/focus.ts";
 
-const subscribeToFocus = (onStoreChange: () => void) => {
-	window.addEventListener("focusin", onStoreChange);
-	window.addEventListener("focusout", onStoreChange);
-
-	return () => {
-		window.removeEventListener("focusin", onStoreChange);
-		window.removeEventListener("focusout", onStoreChange);
-	};
-};
-
-const getFocusedProjectPanel = () =>
-	(document.activeElement?.closest("[data-panel]")?.id as PanelType | undefined) ?? null;
+const getFocusedProjectPanel = (activeElement: Element | null) =>
+	(activeElement?.closest("[data-panel]")?.id as PanelType | undefined) ?? null;
 
 export const useFocusedProjectPanel = (): PanelType | null => {
-	const getSnapshot = () => getFocusedProjectPanel();
-	return useSyncExternalStore(subscribeToFocus, getSnapshot, () => null);
+	const activeElement = useActiveElement();
+	return getFocusedProjectPanel(activeElement);
 };
 
 export const useProjectPanelFocusManager = () => {
@@ -41,7 +32,7 @@ export const useProjectPanelFocusManager = () => {
 		panelElementsRef.current.get(panel)?.focus({ focusVisible: false });
 	};
 	const focusAdjacentPanel = (offset: -1 | 1) => {
-		const currentPanel = getFocusedProjectPanel();
+		const currentPanel = getFocusedProjectPanel(document.activeElement);
 		if (currentPanel === null) return;
 		const nextPanel = orderedPanels[orderedPanels.indexOf(currentPanel) + offset];
 		if (nextPanel === undefined) return;
